style: extract link animation to global css
